refactor(TabList): 모듈화하여 constants 경로로 이동
